+++ Assets/Scripts/Runtime/CH1/Pacmom/RapleyMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 80ad3441a3dad4f4f9193ef2711ef6dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
